Add files via upload
Refacto, clean, and coment
